pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "reporadar"
version = "1.0.0"
description = "Monitor arXiv for papers relevant to your repo and produce a weekly Markdown digest."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.3",
    "jinja2>=3.1",
    "arxiv>=2.0",
]

[project.scripts]
rr = "reporadar.cli:cli"

[project.optional-dependencies]
embeddings = [
    "sentence-transformers>=2.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.hatch.build.targets.wheel]
packages = ["src/reporadar"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["arxiv", "sklearn.*"]
ignore_missing_imports = true
